Last update with float implementation

+++ model.h
@@ -3,7 +3,7 @@
 #define MODEL_H
 #include <stddef.h>
 
-typedef long double real;
+typedef float real;
 typedef enum
 {
     false,
